+++ plain.c
@@ -14,7 +14,7 @@ bool Generate(Particle *buffer, void *context)
     return true;
 }
 
-bool Find(Particle *input, Particle **output, void *context, size_t *sum)
+bool Find(Particle *input, Particle **output, void *context, unsigned int *sum)
 {
     *sum = 0;
     Vec3 center = CENTER;
